Modernal Bot/requirements.txt
discord.py
oauth2
requests